# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: SmartZim Learning Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /exam-dates:
    get:
      operationId: listExamDates
      tags: [examPrep]
      responses:
        "200":
          description: Exam dates
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ExamDate"
    post:
      operationId: createExamDate
      tags: [examPrep]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateExamDateInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ExamDate"
  /exam-dates/{id}:
    delete:
      operationId: deleteExamDate
      tags: [examPrep]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
...
  /channels:
    get:
      operationId: listMyChannels
      tags: [channels]
      responses:
        "200":
          description: My channels
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Channel"
    post:
      operationId: createChannel
      tags: [channels]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateChannelInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Channel"
  /channels/{id}:
    get:
      operationId: getChannel
      tags: [channels]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Channel detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ChannelDetail"
  /channels/{id}/join:
    post:
      operationId: joinChannel
      tags: [channels]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Joined
          content:
            application/json:
              schema:
                type: object
                properties:
                  joined: { type: boolean }
  /channels/{id}/messages:
    get:
      operationId: listChannelMessages
      tags: [channels]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
        - name: limit
          in: query
          schema: { type: string }
        - name: before
          in: query
          schema: { type: string }
      responses:
        "200":
          description: Messages
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ChannelMessage"
    post:
      operationId: sendChannelMessage
      tags: [channels]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SendChannelMessageInput"
      responses:
        "201":
          description: Message sent
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ChannelMessage"
  /channels/{id}/messages/{msgId}:
    delete:
      operationId: deleteChannelMessage
      tags: [channels]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
        - name: msgId
          in: path
          required: true
          schema: { type: integer }
      responses:
        "204":
          description: Deleted
...
  /dm/{userId}:
    get:
      operationId: getDmThread
      tags: [channels]
      parameters:
        - name: userId
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Messages
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/DirectMessage"
    post:
      operationId: sendDirectMessage
      tags: [channels]
      parameters:
        - name: userId
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SendDirectMessageInput"
      responses:
        "201":
          description: Sent
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DirectMessage"
components:
  schemas:
...
    ExamDate:
      type: object
      properties:
        id: { type: integer }
        studentId: { type: integer }
        curriculum: { type: string }
        subject: { type: string }
        paper: { type: ["string", "null"] }
        examDate: { type: string }
        createdAt: { type: string, format: date-time }
      required: [id, studentId, curriculum, subject, examDate, createdAt]
    CreateExamDateInput:
      type: object
      properties:
        subject: { type: string }
        paper: { type: ["string", "null"] }
        examDate: { type: string }
        curriculum: { type: ["string", "null"] }
      required: [subject, examDate]
...
    SendChannelMessageInput:
      type: object
      properties:
        content: { type: string }
        messageType: { type: string }
        replyToId: { type: integer }
      required: [content]
    SendDirectMessageInput:
      type: object
      properties:
        content: { type: string }
      required: [content]
...
    Channel:
      type: object
      properties:
        id: { type: integer }
        name: { type: string }
        school: { type: string }
        grade: { type: ["string", "null"] }
        channelType: { type: string }
        subject: { type: ["string", "null"] }
        description: { type: ["string", "null"] }
        membersCount: { type: integer }
        createdAt: { type: string, format: date-time }
        myRole: { type: ["string", "null"] }
      required: [id, name, school, channelType, membersCount, createdAt]
    ChannelDetail:
      type: object
      properties:
        id: { type: integer }
        name: { type: string }
        school: { type: string }
        grade: { type: ["string", "null"] }
        channelType: { type: string }
        subject: { type: ["string", "null"] }
        description: { type: ["string", "null"] }
        membersCount: { type: integer }
        createdAt: { type: string, format: date-time }
        members:
          type: array
          items:
            type: object
            properties:
              id: { type: integer }
              name: { type: string }
              role: { type: string }
              profilePhotoUrl: { type: ["string", "null"] }
            required: [id, name, role]
      required: [id, name, school, channelType, membersCount, createdAt, members]
    CreateChannelInput:
      type: object
      properties:
        name: { type: string }
        school: { type: string }
        grade: { type: string }
        channelType: { type: string }
        subject: { type: string }
        description: { type: string }
      required: [name, school]
    ChannelMessage:
      type: object
      properties:
        id: { type: integer }
        channelId: { type: integer }
        senderId: { type: integer }
        senderName: { type: ["string", "null"] }
        senderPhoto: { type: ["string", "null"] }
        senderRole: { type: ["string", "null"] }
        content: { type: string }
        messageType: { type: string }
        fileUrl: { type: ["string", "null"] }
        isPinned: { type: boolean }
        replyToId: { type: ["integer", "null"] }
        isDeleted: { type: boolean }
        createdAt: { type: string, format: date-time }
        editedAt: { type: ["string", "null"], format: date-time }
      required: [id, channelId, senderId, content, messageType, isPinned, isDeleted, createdAt]
    DirectMessage:
      type: object
      properties:
        id: { type: integer }
        senderId: { type: integer }
        recipientId: { type: integer }
        content: { type: string }
        fileUrl: { type: ["string", "null"] }
        isRead: { type: boolean }
        createdAt: { type: string, format: date-time }
      required: [id, senderId, recipientId, content, isRead, createdAt]
...